Dto의 위도/경도 타입 변환(Big Decimal -> double)

+++ src/main/java/iriro/saferoute/dto/RiskPointDto.java
@@ -13,7 +13,7 @@ import java.math.BigDecimal;
 @Builder
 public class RiskPointDto {
     private String roadType; //길, 로, 대로
-    private BigDecimal latitude; // 위도
-    private BigDecimal longitude; // 경도
+    private Double latitude; // 위도
+    private Double longitude; // 경도
     private int riskCount; // 같은 도로의 범죄자 수(기본이 1)
 }
