+++ src/main/java/iriro/saferoute/dto/RouteResponseDto.java
@@ -13,10 +13,10 @@ import java.util.List;
 @Data
 @Builder
 public class RouteResponseDto {
-    private BigDecimal start_latitude;
-    private BigDecimal start_longitude;
-    private BigDecimal end_latitude;
-    private BigDecimal end_longitude;
+    private Double start_latitude;
+    private Double start_longitude;
+    private Double end_latitude;
+    private Double end_longitude;
 
     private Integer totalTime;
     private Integer totalDistance;
