+++ src/main/java/iriro/saferoute/dto/SafetyFacPointDto.java
@@ -11,6 +11,6 @@ import java.math.BigDecimal;
 public class SafetyFacPointDto {
     private String facType; // 안전시설물 타입
     private int safeCount; // 안전 시설물 개수
-    private BigDecimal latitude; // 위도
-    private BigDecimal longitude; // 경도
+    private Double latitude; // 위도
+    private Double longitude; // 경도
 }
